Place \asAuthor above ossia instead of before
Placing the \asAuthor text in the “instrument name” before the ossia
staff only works if that staff is not at the beginning of a line, and I
intend to leave the line breaking completely to LilyPond, so we have no
idea when the line will break and when we could use the instrument name.
Instead, always place the instruction above the staff, as the original
score does in situations where the ossia are at the beginning of the
line (pages 6, 8, 15).